requirements.txt
boto3==1.34.69
botocore==1.34.69
ncloud-sdk==1.1.8
python-dotenv==1.0.1
requests==2.31.0 